Move icloud services to separate module
